+++ src/main/java/org/opensearch/knn/index/query/PerLeafResult.java
@@ -72,6 +72,11 @@ public class PerLeafResult {
         }
 
         @Override
+        public int nextClearBit(int start, int end) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
         public long ramBytesUsed() {
             throw new UnsupportedOperationException();
         }
@@ -129,6 +134,11 @@ public class PerLeafResult {
 
         @Override
         public int nextSetBit(int start, int end) {
+            throw new UnsupportedOperationException();
+        }
+
+        @Override
+        public int nextClearBit(int start, int end) {
             throw new UnsupportedOperationException();
         }