+++ src/test/java/org/opensearch/knn/index/codec/KNN1040Codec/KNN1040HnswScalarQuantizedVectorsFormatTests.java
@@ -5,13 +5,13 @@
 
 package org.opensearch.knn.index.codec.KNN1040Codec;
 
-import org.apache.lucene.codecs.lucene104.Lucene104ScalarQuantizedVectorsFormat.ScalarEncoding;
+import org.apache.lucene.util.quantization.QuantizedByteVectorValues.ScalarEncoding;
 import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.index.engine.BuiltinKNNEngine;
 
 import java.util.concurrent.Executors;
 
-import static org.apache.lucene.codecs.lucene104.Lucene104ScalarQuantizedVectorsFormat.ScalarEncoding.SINGLE_BIT_QUERY_NIBBLE;
+import static org.apache.lucene.util.quantization.QuantizedByteVectorValues.ScalarEncoding.SINGLE_BIT_QUERY_NIBBLE;
 
 public class KNN1040HnswScalarQuantizedVectorsFormatTests extends KNNTestCase {